reworked half of the sim
