--- TwoDimension/TwoDimension/LoadedCapillaries/BandwidthConverter.py
@@ -1,8 +1,0 @@
-import numpy as np
-
-df = 10
-cWL = 1.55
-c = 1
-
-
-print(df*cWL**2/c)
