--- TwoDimension/TwoDimension/SidePolishedFibre.py
@@ -1,1 +1,0 @@
-
