--- TwoDimension/TwoDimension/WGMs.py
@@ -1,1 +1,0 @@
-
